фикс замечаний из ПР
