+++ ImageFeed/Auth/WebViewPresenter.swift
@@ -11,7 +11,7 @@ final class WebViewPresenter: WebViewPresenterProtocol {
    
     weak var view: WebViewControllerProtocol?
     
-    var authHelper: AuthHelperProtocol
+    private let authHelper: AuthHelperProtocol
     
     init(authHelper: AuthHelperProtocol) {
         self.authHelper = authHelper
