+++ ImageFeed/SplashViewController/SplashViewController.swift
@@ -127,4 +127,3 @@ extension SplashViewController {
         }
     }
 }
-
