+++ ImageFeed/TabBar/ImagesList/ImagesListCell.swift
@@ -30,4 +30,3 @@ final class ImagesListCell: UITableViewCell {
         }
     }
 }
-
