+++ ImageFeed/TabBar/Profile/LogoutService.swift
@@ -37,5 +37,3 @@ final class LogoutService {
         profileViewController.cleanProfileData()
     }
 }
-    
-
