+++ ImageFeed/TabBar/Profile/ProfileImageResponseBody.swift
@@ -24,4 +24,3 @@ struct UserResult: Decodable {
     }
     
 }
-
