+++ ImageFeed/TabBar/Profile/ProfileResultResponseBody.swift
@@ -5,7 +5,6 @@ struct ProfileResultResponseBody: Codable {
     let firstName: String
     let lastName: String
     let bio: String?
-
     
     private enum CodingKeys: String, CodingKey {
         case userName = "username"
@@ -14,17 +13,3 @@ struct ProfileResultResponseBody: Codable {
         case bio
     }
 }
-
-//struct Profile: Codable{
-//    let username: String
-//    let name: String
-//    let loginName: String
-//    let bio: String?
-//    
-//}
-
-//struct ProfileImage: Codable {
-//    let small: String
-//    let medium: String
-//    let large: String
-//}
