+++ ImageFeed/TabBar/Profile/ProfileViewController.swift
@@ -178,4 +178,3 @@ final class ProfileViewController:UIViewController, ProfileViewControllerProtoco
                                     + logoutButtonConstraints)
     }
 }
-
