+++ ImageFeed/URL/URLSessionExtension.swift
@@ -34,4 +34,3 @@ extension URLSession {
         return task
     }
 }
-
