+++ ImageFeedTests/WebViewTests/WebViewPresenterSpy.swift
@@ -17,4 +17,3 @@ final class WebViewPresenterSpy: WebViewPresenterProtocol {
         return nil
     }
 }
-
